+++ src/qcoder/__init__.py
@@ -1,3 +1,3 @@
 __all__ = []
-__version__ = "0.6.0a1"
+__version__ = "0.6.0a2"
 file = __file__